backend/requirements-test.txt
httpx>=0.27.0
pytest>=8.3.0
pytest-html>=4.1.0
allure-pytest>=2.13.5
faker>=24.0.0
python-dotenv>=1.0.0
